requirements.txt
asn1crypto==0.24.0
awacs==0.9.7
awscli==1.17.10
AWSScout2==3.2.1
boto3==1.11.10
botocore==1.14.10
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.1.4
docutils==0.15.2
enum34==1.1.6
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
futures==3.3.0
iampoliciesgonewild==1.0.6.2
idna==2.8
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
MySQL-python==1.2.5
netaddr==0.7.19
opinel==3.3.4
pyasn1==0.4.8
pycrypto==2.6.1
pygobject==3.26.1
python-dateutil==2.8.1
pkg-resources==0.0.0
pyxdg==0.25
PyYAML==5.3
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.2
SecretStorage==2.3.1
six==1.14.0
SQLAlchemy==1.3.13
troposphere==2.5.3
urllib3==1.25.8
Werkzeug==1.0.0
mysql-connector-python
